apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-synthesia-api"
version = "0.1.0"
description = "Synesthesia AI API Gateway"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "sqlalchemy==2.0.23",
    "alembic==1.13.0",
    "asyncpg==0.29.0",
    "redis==5.0.1",
    "boto3==1.34.0",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "httpx==0.25.2",
    "opentelemetry-api==1.21.0",
    "opentelemetry-sdk==1.21.0",
    "opentelemetry-instrumentation-fastapi==0.42b0",
    "opentelemetry-instrumentation-sqlalchemy==0.42b0",
    "prometheus-client==0.19.0",
    "sentry-sdk[fastapi]==1.38.0",
    "crewai==0.22.5",
    "langchain==0.0.350",
    "langchain-openai==0.0.2",
    "pgvector==0.2.4"
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "black==23.11.0",
    "isort==5.12.0",
    "mypy==1.7.1",
    "ruff==0.1.6"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
line-length = 88
target-version = "py311"
